ADD parse_bullet_list
Add logic to parse unordered lists. This only works if if the list is at
top level, I'm not sure yet how I will handle mixed typed lists, like
embedding a unordered list in a ordered list, that would be very
different syntax in wiki markup and in markdown. That will be a good
late task, when most commun use case are already covered, if I feel like
it.