ADD <nowiki> support
Allow to bypass parsing wiki markup.

There's a catch, though. The wiki markup is meant to be exported to
html. That means that if you add newlines inside a nowiki text content,
it will make a single line, because html doesn't recognise "\n" as a tag
or having any special meaning, it's just a space, for it. But it's not
the case in markdown. So, to keep the output matching the intent of the
user writing the wiki code, we must convert newlines into space when
inside a nowiki tag.

This also has other implications, like `*text*` inside a nowiki should
not be parsed as markdown, but this is an edge case, I'll see if it
actually happens before implementing something to handle it.