REFACTOR convertion of definition list
Actually, using html <dl> for definition list won't do, since markdown
doesn't support mixing html and mardown : if something starts with html
code, then it's html. This means that if a definition list contains a
link, it won't be parsed, unless it's html. And that would suddenly
require to implement a html dumper as well, let's not go that way. :)

Instead, a definition list is now a bold paragraph for the term,
following by a bullet list for the definitions.

I needed to add a fourth parameter to the parameters of dumping node
handlers, so there are a lot of changes, to use the same parameter
struct pattern than in other files.